+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added 
  - Added documentation on Writing custom HealthIndicators
+ - Added ``management.health.defaults.enabled`` to allow disabling all HealthIndicators
 ### Changed
  - Changed the /health endpoint, it now reports custom health indicators using thier bean names less the 'HealthIndicator' suffix (if it exists)
 ### Deprecated 
 ### Removed
 ### Fixed
+ - Fixed ISSUE-2: Add management.health.defaults.enabled property
  - Fixed ISSUE-1: HealthIndicator names in /health
 ### Security